Add validation to user with annotations. Refactor 1 SQL query. Fix 1 bug with id (now long instead of int, and that's why appears 1 bug). Fix bug with no unique result when create order

+++ src/main/java/com/epam/esm/dao/impl/OrderDaoImpl.java
@@ -64,7 +64,7 @@ public class OrderDaoImpl implements OrderDao {
         String query = GET_ORDER_BY_USER_ID +
                 "order by " + searchCriteria.getSortBy() + " " + searchCriteria.getSortType();
         TypedQuery<Order> typedQuery = entityManager.createQuery(query, Order.class);
-        typedQuery.setParameter("userId", id);
+        typedQuery.setParameter("userId", (long)id);
         typedQuery.setFirstResult((page - 1) * size);
         typedQuery.setMaxResults(size);
